@@ -1,6 +1,13 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css"
+import { createRoot } from "react-dom/client";
+import "./index.css";
 import App from "./App";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const rootElement = document.getElementById("root");
+const root = createRoot(rootElement);
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
+
