@@ -1,11 +1,37 @@
-import React from 'react'
-
+import React from "react";
+import { Link, NavLink } from "react-router-dom";
+import { SiShopware } from "react-icons/si";
+import { MdOutlineCancel } from "react-icons/md";
+import { TooltipComponent } from "@syncfusion/ej2-react-popups";
+import { links } from "../data/dummy";
 const SideBar = () => {
+  const activeMenu = true;
   return (
-    <div>
-      
+    <div className="ml-3 h-screen md:overflow-auto overflow-auto md:hover:overflow-auto pb-10">
+      {activeMenu && (
+        <>
+          <div className="flex justify-between items-center">
+            <Link
+              to="/"
+              onClick={() => {}}
+              className="items-center gap-3 ml-3 mt-4 flex text-xl font-extrabold tracking-tight dark:text-white text-slate-900"
+            >
+              <SiShopware /> <span>Shoppy</span>
+            </Link>
+            <TooltipComponent content="Menu" position="BottomCenter">
+              <button
+                type="button"
+                onClick={() => {}}
+                className="text-xl rounded-full p-3 hover:bg-light-gray mt-4 block md:hidden"
+              >
+                <MdOutlineCancel />
+              </button>
+            </TooltipComponent>
+          </div>
+        </>
+      )}
     </div>
-  )
-}
+  );
+};
 
-export default SideBar
+export default SideBar;
