@@ -3,7 +3,7 @@ import React from 'react'
 const Ecommerce = () => {
   return (
     <div>
-      
+      welcome 
     </div>
   )
 }
